Use Application.compile_env for platforms to prevent re-running without recompiling dependency when config changes
Otherwise, in a client of the package, updating the config silently
fails to take effect until the dependency is recompiled, since the
config value is read at compile-time.

With this, changing the config and trying to re-run will result in this
error:

    ERROR! the application :live_view_native has a different value set for key :platforms during runtime compared to compile time. Since this application environment entry was marked as compile time, this difference can lead to different behaviour than expected:

until the dependency is recompiled so the new config change takes
effect.