Fixed new template re-compilation issue
If a new template was added when the Phoenix server is running that new template
wasn't being picked up by the CodeReloader because it wasn't being tracked. The originaly implementation
copy-pasta'd the Phoenix LiveView implementation which gets around this because the template
names are known AOT even if the template doesn't exist yet so that `@external_resource` was tracked
but in LVN because we incorporate the TARGET into the template name (i.e. `home_live.swiftui+ios.neex`)
we cannot use this method. The solution was to implement `__mix_recompile__?` and track file collection
hashes for comparing if a new file was added or not that *should* be treated as a template for the
given LiveView Native Render Component